requirements.txt
streamlit==1.33.0
ollama>=0.4.8
python-docx            # for reading .docx resumes
PyMuPDF                # for reading .pdf resumes (imported as fitz)
